feat(staking): change BASE_VALUE to 10**28 and change interest type t…

+++ workspace/apps/staking/contracts/src/types.cairo
@@ -3,5 +3,5 @@ pub type Commission = u16;
 pub type TimeDelta = u64;
 pub type TimeStamp = u64;
 pub type Amount = u128;
-pub type Index = u64;
+pub type Index = u128;
 pub type Inflation = u16;
